+++ src/core/components/index.ts
@@ -17,6 +17,7 @@ export { DefaultChaiBlocks } from "../components/sidepanels/panels/add-blocks/De
 export { DarkMode as DarkModeSwitcher } from "../components/canvas/topbar/DarkMode.tsx";
 export { Breakpoints as ScreenSizes } from "../components/canvas/topbar/Breakpoints.tsx";
 export { UndoRedo } from "../components/canvas/topbar/UndoRedo.tsx";
+export { default as JSONFormFieldTemplate } from "./settings/JSONFormFieldTemplate.tsx";
 export {
   ChaiBuilderCanvas,
   ThemeOptions,
